Add 3D phone model and scene
